refactor: extract code about dashboard serialization to its own file
